Merge

+++ src/main/java/org/example/practiceproject1/models/HockeyPlayerRepository.java
@@ -2,5 +2,5 @@ package org.example.practiceproject1.models;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface HockeyPlayerRepository extends CrudRepository<HockeyPlayer, Integer> {
+public interface HockeyPlayerRepository extends CrudRepository<HockeyPlayer, Long> {
 }
